requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
langchain
langchain-core
langchain-community
langchain-text-splitters
langchain-google-genai
google-genai
langchain_chroma
tqdm
langchain-elasticsearch